requirements.txt
Django==3.0.1
django-filter==2.2.0
djangorestframework==3.11.0
